feat(profiles): make walking paths on squares an opt-in flag in foot.lua
Meshing open areas meant reaching for a second profile. foot_area.lua was
a copy of foot.lua with the area calls added, and copies drift: it had
fallen behind on sidewalk=separate handling, foot=use_sidepath, the
audible-fence exception, motorroad and max_collapse_distance. Anyone
routing over plazas was quietly getting an older pedestrian profile.

Put the feature in foot.lua behind a flag that defaults to off, and let
foot_area.lua turn it on. The flag reads an optional global before it
shadows it, so requiring the profile with the global already set enables
meshing without editing the file, and foot_area.lua becomes the two
lines that do exactly that. profiles/test.lua already requires car.lua
the same way.

Registering plazas mapped as multipolygon relations needs
process_relation, which the extractor only calls at api_version 4, so
foot.lua moves from 2 to 4. The only incompatibility is the turn
function: api_version 4 drops direction_modifier, which up to 3 was
synthesised as is_u_turn ? UTurn : Straight, so the check becomes
turn.is_u_turn and asks the same question. Everything else v4 adds is
additive.

With the flag off the extractor writes no .osrm.openareas files and the
profile behaves as before. The whole features/foot suite passes either
way, and foot_area.lua still extracts monaco in library-extract-tests.

Relates to #7683
