+++ profiles/foot.lua
@@ -1,14 +1,26 @@
 -- Foot profile
 
-api_version = 2
+api_version = 4
 
 Set = require('lib/set')
 Sequence = require('lib/sequence')
 Handlers = require("lib/way_handlers")
 find_access_tag = require("lib/access").find_access_tag
 
+-- Walking paths across the interior of open areas such as plazas and squares.
+-- Off by default: it makes extraction slower and the generated ways are
+-- synthetic, so it is opted into rather than assumed. Set this to true, or set
+-- the global before requiring this profile the way profiles/foot_area.lua does.
+-- See docs/areas.md.
+local enable_area_meshing = enable_area_meshing or false
+
 function setup()
   local walking_speed = 5
+
+  if enable_area_meshing then
+    area_manager:init('visgraph+dijkstra')
+  end
+
   return {
     properties = {
       weight_name                   = 'duration',
@@ -139,7 +151,7 @@ function setup()
   }
 end
 
-function process_node(profile, node, result)
+function process_node(profile, node, result, relations)
   -- parse access and barrier tags
   local access = find_access_tag(node, profile.access_tags_hierarchy)
   if access then
@@ -203,7 +215,7 @@ local function handle_sidewalk_separate(profile, way, result, data)
 end
 
 -- main entry point for processsing a way
-function process_way(profile, way, result)
+function process_way(profile, way, result, relations)
   -- the intial filtering of ways based on presence of tags
   -- affects processing times significantly, because all ways
   -- have to be checked.
@@ -233,6 +245,12 @@ function process_way(profile, way, result)
   -- of the prefetched tags to be present, ie. the data table
   -- cannot be empty
   if next(data) == nil then     -- is the data table empty?
+    return
+  end
+
+  -- Hand the area to the mesher instead of routing along its outline.
+  if enable_area_meshing and data.highway == 'pedestrian' and way:has_true_tag('area') then
+    area_manager:way(way)
     return
   end
 
@@ -288,10 +306,18 @@ function process_way(profile, way, result)
   WayHandlers.run(profile, way, result, data, handlers)
 end
 
+-- Plazas mapped as multipolygon relations rather than closed ways.
+function process_relation(profile, relation)
+  if enable_area_meshing and relation:has_tag('type', 'multipolygon')
+     and relation:has_tag('highway', 'pedestrian') then
+    area_manager:relation(relation)
+  end
+end
+
 function process_turn (profile, turn)
   turn.duration = 0.
 
-  if turn.direction_modifier == direction_modifier.u_turn then
+  if turn.is_u_turn then
      turn.duration = turn.duration + profile.properties.u_turn_penalty
   end
 
@@ -310,5 +336,6 @@ return {
   setup = setup,
   process_way =  process_way,
   process_node = process_node,
-  process_turn = process_turn
+  process_turn = process_turn,
+  process_relation = process_relation
 }
